Address most comments in the review
