+++ easybuild/framework/easyconfig/tools.py
@@ -728,3 +728,21 @@ def avail_easyblocks():
                                        easyblock_mod_name, easyblocks[easyblock_mod_name]['loc'], path)
 
     return easyblocks
+
+
+def copy_ecs_to_target(determined_paths, target_path, prefix=False):
+    """
+    Copy list of easyconfigs to specified path
+
+    :param determined_paths: paths to ecs to copy
+    :param target_path: target to copy files to
+    :param prefix: include message prefix characters
+    """
+    if len(determined_paths) == 1:
+        copy_file(determined_paths[0], target_path)
+        print_msg("%s copied to %s" % (os.path.basename(determined_paths[0]), target_path), prefix=prefix)
+    elif len(determined_paths) > 1:
+        copy_files(determined_paths, target_path)
+        print_msg("%d file(s) copied to %s" % (len(determined_paths), target_path), prefix=prefix)
+    else:
+        raise EasyBuildError("One of more files to copy should be specified!")
